+++ src/main/java/org/gotchafish/spot/dao/UserSpotDAO.java
@@ -16,7 +16,7 @@ public interface UserSpotDAO {
      * @return 성공하면 true, 실패하면 false
      * @throws SQLException DB 처리 중 예외가 발생한 경우
      */
-    boolean insert(Connection conn, Long userId, int spotId) throws SQLException;
+    boolean insert(Connection conn, Long userId, Long spotId) throws SQLException;
 
     /**
      * 회원이 특정 낚시터를 이미 잠금 해제했는지 확인한다.
@@ -29,7 +29,7 @@ public interface UserSpotDAO {
      * @return 이미 잠금 해제했으면 true, 아니면 false
      * @throws SQLException DB 처리 중 예외가 발생한 경우
      */
-    boolean existsByUserIdAndSpotId(Connection conn, Long userId, int spotId) throws SQLException;
+    boolean existsByUserIdAndSpotId(Connection conn, Long userId, Long spotId) throws SQLException;
 
     /**
      * 회원이 잠금 해제한 낚시터 ID 목록을 조회한다.
@@ -42,5 +42,5 @@ public interface UserSpotDAO {
      * @return 잠금 해제한 낚시터 ID 목록
      * @throws SQLException DB 처리 중 예외가 발생한 경우
      */
-    List<Integer> findSpotIdsByUserId(Connection conn, Long userId) throws SQLException;
+    List<Long> findSpotIdsByUserId(Connection conn, Long userId) throws SQLException;
 }
